% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alternate_priors.R
\name{alternate_priors}
\alias{alternate_priors}
\title{Function \code{alternate_priors}}
\usage{
alternate_priors()
}
\value{
string vector of names of alternate priors
}
\description{
Function \code{alternate_priors}
}
\details{
alternate priors for phi_g, alpha_g, and delta_g
}

