% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assign_samplers.R
\name{assign_samplers}
\alias{assign_samplers}
\title{Function \code{assign_samplers}}
\usage{
assign_samplers(chain, configs)
}
\arguments{
\item{chain}{\code{Chain} object}

\item{configs}{\code{Configs} object}
}
\value{
vector \code{Chain} object with samplers assigned
}
\description{
Assigns the right samplers to the "Sampler" slots of a \code{Chain} object
}

