% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{backends}
\alias{backends}
\title{Function \code{backends}}
\usage{
backends()
}
\value{
Character vector of packages.
}
\description{
Character vector of available MCMC backends for \code{fbseq}.
The elements are the R packages containing the backends, and the names
are the corresponding \code{backend} arguments to the \code{fbseq} function.
}

