% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_gelman_mcmc.R
\name{calc_gelman}
\alias{calc_gelman}
\title{Function \code{calc_gelman}}
\usage{
calc_gelman(chain_list)
}
\arguments{
\item{chain_list}{list of \code{Chain} objects}
}
\value{
Gelman potential scale reduction factors on a list of parallel independent chains.
}
\description{
calculate Gelman potential scale reduction factors on a list of parallel independent chains.
}

