% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chain.R
\docType{class}
\name{Chain-class}
\alias{Chain-class}
\title{Class \code{Chain}}
\description{
the main storage object for a heterosis MCMC chain.
Create a new one with a call to \code{Chain()} and run MCMC
by feeding one into \code{single_chain()}.
}
\section{Slots}{

\describe{
\item{\code{diag}}{convergence diagnostic to use. Can be "gelman" or "none".}

\item{\code{ess}}{Minimum effective sample size for all parameters}

\item{\code{psrf_tol}}{upper threshold for Gelman-Rubin potential scale reduction factors (if diag is "gelman")}

\item{\code{max_attempts}}{Maximum number of retries for assessing convergence and generating enough effective samples.
Can be set to Inf to run indefinitely.}

\item{\code{nchains_diag}}{number of independent chains to run (including this one) to use
convergence diagnostics that require multiple chains.}

\item{\code{psrf}}{Gelman-Rubin potential scale reduction factors, if running \code{run_gelman_mcmc()}}

\item{\code{counts}}{vector of RNA-seq read counts. Type \code{matrix(chain@counts, ncol = chain@N)}
to recover the original RNA-seq count data matrix with features/genes
as rows and samples/libraries as columns with .}

\item{\code{group}}{Experimental design. A vector of integers,
one for each RNA-seq sample/library, denoting the genetic
variety of that sample. You must use 1 for parent 1, 2 for parent 2,
and 3 for the hybrid.}

\item{\code{sums_n}}{Sample/library-specific sums of counts.}

\item{\code{sums_g}}{Feature/gene-specific sums of counts.}

\item{\code{returns}}{named integer vector of 0/1 values naming
the variables whose MCMC samples you want to return.}

\item{\code{updates}}{named integer vector of 0/1 values naming
the variables to compute/update during the MCMC.}

\item{\code{samples_return}}{indices of RNA-seq samples/libraries whose parameter samples you want to return.}

\item{\code{features_return}}{indices of features/genes whose parameter samples you want to return.}

\item{\code{samples_return_eps}}{Indices of RNA-seq samples/libraries n for which epsilon_{n, g} is updated/returned.}

\item{\code{features_return_eps}}{Indices of features/genes g for which epsilon_{n, g} is updated/returned.}

\item{\code{M}}{number of MCMC iterations (not including burnin or thinning)}

\item{\code{N}}{number of RNA-seq samples/libraries}

\item{\code{G}}{number of features/genes}

\item{\code{Nreturn}}{length of slot \code{samples_return}.}

\item{\code{Greturn}}{length of slot \code{features_return}.}

\item{\code{NreturnEps}}{length of slot \code{samples_return_eps}.}

\item{\code{GreturnEps}}{length of slot \code{features_return_eps}.}

\item{\code{burnin}}{MCMC burnin, the number of MCMC iterations to ignore at the beginning of each chain}

\item{\code{mphtol}}{tolerance theshold for mid-parent heterosis (on a log base e scale)}

\item{\code{thin}}{MCMC thinning interval, number of iterations to skip in between iterations to return.}

\item{\code{verbose}}{number of times to print out progress during burnin and the actual MCMC.
If \code{verbose} > 0, then progress messages will also print during setup and cleanup.}

\item{\code{seeds}}{Random number generator seeds.}

\item{\code{hph}}{gene/feature-specific high-parent heterosis probabilities. The order of the
probabilities respects the ordering of features/genes in the original RNA-seq dataset.}

\item{\code{lph}}{gene/feature-specific low-parent heterosis probabilities. The order of the
probabilities respects the ordering of features/genes in the original RNA-seq dataset.}

\item{\code{mph}}{gene/feature-specific mid-parent heterosis probabilities. The order of the
probabilities respects the ordering of features/genes in the original RNA-seq dataset.}

\item{\code{phiPrior}}{Integer indicating the family of priors on the phi_g's after integrating out the xi_phi's.
The integer is either 0 for normal or the index of the corresponding prior in the
return value of \code{alternate_priors()}.}

\item{\code{alpPrior}}{Integer indicating the family of priors on the alp_g's after integrating out the xi_alp's.
The integer is either 0 for normal or the index of the corresponding prior in the
return value of \code{alternate_priors()}.}

\item{\code{delPrior}}{Integer indicating the family of priors on the del_g's after integrating out the xi_del's.
The integer is either 0 for normal or the index of the corresponding prior in the
return value of \code{alternate_priors()}.}

\item{\code{dRho}}{initialization constant}

\item{\code{aRho}}{initialization constant}

\item{\code{bRho}}{initialization constant}

\item{\code{dGam}}{initialization constant}

\item{\code{aGam}}{initialization constant}

\item{\code{bGam}}{initialization constant}

\item{\code{cPhi}}{initialization constant}

\item{\code{cAlp}}{initialization constant}

\item{\code{cDel}}{initialization constant}

\item{\code{sPhi}}{initialization constant}

\item{\code{sAlp}}{initialization constant}

\item{\code{sDel}}{initialization constant}

\item{\code{kPhi}}{initialization constant}

\item{\code{kAlp}}{initialization constant}

\item{\code{kDel}}{initialization constant}

\item{\code{rPhi}}{initialization constant}

\item{\code{rAlp}}{initialization constant}

\item{\code{rDel}}{initialization constant}

\item{\code{nuRho}}{vector of model parameter samples}

\item{\code{nuGam}}{vector of model parameter samples}

\item{\code{tauRho}}{vector of model parameter samples}

\item{\code{tauGam}}{vector of model parameter samples}

\item{\code{thePhi}}{vector of model parameter samples}

\item{\code{theAlp}}{vector of model parameter samples}

\item{\code{theDel}}{vector of model parameter samples}

\item{\code{sigPhi}}{vector of model parameter samples}

\item{\code{sigAlp}}{vector of model parameter samples}

\item{\code{sigDel}}{vector of model parameter samples}

\item{\code{phi}}{vector of model parameter samples}

\item{\code{alp}}{vector of model parameter samples}

\item{\code{del}}{vector of model parameter samples}

\item{\code{rho}}{vector of model parameter samples}

\item{\code{gam}}{vector of model parameter samples}

\item{\code{xiPhi}}{vector of model parameter samples}

\item{\code{xiAlp}}{vector of model parameter samples}

\item{\code{xiDel}}{vector of model parameter samples}

\item{\code{eps}}{vector of model parameter samples (flattened from a 2D array)}

\item{\code{nuRhoStart}}{vector of starting values}

\item{\code{nuGamStart}}{vector of starting values}

\item{\code{tauRhoStart}}{vector of starting values}

\item{\code{tauGamStart}}{vector of starting values}

\item{\code{thePhiStart}}{vector of starting values}

\item{\code{theAlpStart}}{vector of starting values}

\item{\code{theDelStart}}{vector of starting values}

\item{\code{sigPhiStart}}{vector of starting values}

\item{\code{sigAlpStart}}{vector of starting values}

\item{\code{sigDelStart}}{vector of starting values}

\item{\code{phiStart}}{vector of starting values}

\item{\code{alpStart}}{vector of starting values}

\item{\code{delStart}}{vector of starting values}

\item{\code{rhoStart}}{vector of starting values}

\item{\code{gamStart}}{vector of starting values}

\item{\code{xiPhiStart}}{vector of starting values}

\item{\code{xiAlpStart}}{vector of starting values}

\item{\code{xiDelStart}}{vector of starting values}

\item{\code{epsStart}}{vector of starting values (flattened from a 2D array)}

\item{\code{nuRhoPostMean}}{vector of posterior means after burnin}

\item{\code{nuGamPostMean}}{vector of posterior means after burnin}

\item{\code{tauRhoPostMean}}{vector of posterior means after burnin}

\item{\code{tauGamPostMean}}{vector of posterior means after burnin}

\item{\code{thePhiPostMean}}{vector of posterior means after burnin}

\item{\code{theAlpPostMean}}{vector of posterior means after burnin}

\item{\code{theDelPostMean}}{vector of posterior means after burnin}

\item{\code{sigPhiPostMean}}{vector of posterior means after burnin}

\item{\code{sigAlpPostMean}}{vector of posterior means after burnin}

\item{\code{sigDelPostMean}}{vector of posterior means after burnin}

\item{\code{phiPostMean}}{vector of posterior means after burnin}

\item{\code{alpPostMean}}{vector of posterior means after burnin}

\item{\code{delPostMean}}{vector of posterior means after burnin}

\item{\code{rhoPostMean}}{vector of posterior means after burnin}

\item{\code{gamPostMean}}{vector of posterior means after burnin}

\item{\code{xiPhiPostMean}}{vector of posterior means after burnin}

\item{\code{xiAlpPostMean}}{vector of posterior means after burnin}

\item{\code{xiDelPostMean}}{vector of posterior means after burnin}

\item{\code{epsPostMean}}{vector of posterior means after burnin (flattened from a 2D array)}

\item{\code{nuRhoPostMeanSq}}{vector of posterior means of squares after burnin}

\item{\code{nuGamPostMeanSq}}{vector of posterior means of squares after burnin}

\item{\code{tauRhoPostMeanSq}}{vector of posterior means of squares after burnin}

\item{\code{tauGamPostMeanSq}}{vector of posterior means of squares after burnin}

\item{\code{thePhiPostMeanSq}}{vector of posterior means of squares after burnin}

\item{\code{theAlpPostMeanSq}}{vector of posterior means of squares after burnin}

\item{\code{theDelPostMeanSq}}{vector of posterior means of squares after burnin}

\item{\code{sigPhiPostMeanSq}}{vector of posterior means of squares after burnin}

\item{\code{sigAlpPostMeanSq}}{vector of posterior means of squares after burnin}

\item{\code{sigDelPostMeanSq}}{vector of posterior means of squares after burnin}

\item{\code{phiPostMeanSq}}{vector of posterior means of squares after burnin}

\item{\code{alpPostMeanSq}}{vector of posterior means of squares after burnin}

\item{\code{delPostMeanSq}}{vector of posterior means of squares after burnin}

\item{\code{rhoPostMeanSq}}{vector of posterior means of squares after burnin}

\item{\code{gamPostMeanSq}}{vector of posterior means of squares after burnin}

\item{\code{xiPhiPostMeanSq}}{vector of posterior means of squares after burnin}

\item{\code{xiAlpPostMeanSq}}{vector of posterior means of squares after burnin}

\item{\code{xiDelPostMeanSq}}{vector of posterior means of squares after burnin}

\item{\code{epsPostMeanSq}}{vector of posterior means of squares after burnin (flattened from a 2D array)}
}}

