% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_constructor.R
\name{Chain}
\alias{Chain}
\title{Function \code{Chain}}
\usage{
Chain(scenario, configs = Configs(), starts = Starts(), slots = NULL)
}
\arguments{
\item{scenario}{a \code{Scenario} object with count data, the design matrix, etc.}

\item{configs}{A \code{Configs} object of MCMC control parameters.}

\item{starts}{A \code{Starts} object of model parameter starting bounds.}

\item{slots}{a list of slots to be plugged into a \code{Chain} object}
}
\value{
a \code{Chain} object
}
\description{
Construct a \code{Chain} object out of \code{Scenario},
\code{Configs}, and \code{Starts} objects or with a list of slots listed 
by \code{help("Chain-class")}.
}
\details{
If \code{slots} is not \code{NULL}, all other arguments will be dropped.
}
\seealso{
\code{help("Chain-class")}
}

