% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_backend.R
\name{check_backend}
\alias{check_backend}
\title{Function \code{check_backend}}
\usage{
check_backend(backend_package = "fbseqCUDA")
}
\arguments{
\item{backend_package}{defaults to "fbseqCUDA". Other options include
"fbseqSerial", which does not use any parallel computing.}
}
\description{
Checks that a given backend package.
If the package is not installed, the function quits in error.
}

