% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fbseq_input.R
\name{check_fbseq_input}
\alias{check_fbseq_input}
\title{Function \code{check_fbseq_input}}
\usage{
check_fbseq_input(backend, processes = 1, threads = 1)
}
\arguments{
\item{backend}{Backend for parallel computing. Options are
"CUDA" and "OpenMP".}

\item{processes}{number of CPU processes to fork for 
the additional chains. This argument is automatically reset to 1 for non-serial
backends because additional \code{parallel::mclapply} processes interfere
with other modes of parallelism.
For some other backends, chains will be distributed accross processes.}

\item{threads}{Number of threads for the OpenMP implementation.}
}
\value{
Corrected number of processes.
}
\description{
Checks input to the \code{fbseq} function.
}

