% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cuda_usage.R
\name{check_fbseqCUDA}
\alias{check_fbseqCUDA}
\title{Function \code{check_fbseqCUDA}}
\usage{
check_fbseqCUDA()
}
\description{
Checks that package \code{fbseqCUDA} is installed.
If the package is not installed, the function quits in error.
}

