% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cuda_usage.R
\name{check_heterosisCUDA}
\alias{check_heterosisCUDA}
\title{Function \code{check_heterosisCUDA}}
\usage{
check_heterosisCUDA()
}
\description{
Checks that package \code{heterosisCUDA} is installed.
If the package is not installed, the function quits in error.
}

