% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inference.R
\name{check_inference}
\alias{check_inference}
\title{Function \code{check_inference}}
\usage{
check_inference(inference)
}
\arguments{
\item{inference}{a \code{Inference} object}
}
\value{
a \code{Inference} object
}
\description{
Function \code{check_inference}
}
\details{
checks an \code{Inference} object for inconsistencies.
}

