% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{check_scenario}
\alias{check_scenario}
\title{Function \code{check_scenario}}
\usage{
check_scenario(scenario)
}
\arguments{
\item{scenario}{a \code{Scenario} object}
}
\value{
a \code{Scenario} object
}
\description{
checks an \code{Scenario} object for inconsistencies.
}

