% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concatenate.R
\name{concatenate}
\alias{concatenate}
\title{Function \code{concatenate}}
\usage{
concatenate(chain1, chain2)
}
\arguments{
\item{chain1}{a \code{Chain} object}

\item{chain2}{a \code{Chain} object}
}
\value{
a \code{Chain} object
}
\description{
Concatenate 2 \code{Chain} objects that are chronologically adjacent.
}
\details{
If you run \code{chain1} and if \code{chain2 = run_mcmc(chain1)}, then you can call
\code{concatenate(chain1, chain2)} to combine them. The parameter samples
will be combined, the starting values will be the same as for chain2, and
\code{concatenate(chain1, chain2)@M = chain1@M + chain2@M}. Also,
\code{concatenate(chain1, chain2)@hph = (chain1@M * chain1@hph + chain2@M * chain2@hph) / (chain1@M + chain2@M)}
The other heterosis probabilities are similarly combined.
}

