% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/configs.R
\docType{class}
\name{Configs-class}
\alias{Configs-class}
\title{Class \code{Configs}}
\description{
A collection of MCMC control parameters.
}
\section{Slots}{

\describe{
\item{\code{diag}}{convergence diagnostic to use. Can be "gelman" or "none".}

\item{\code{ess}}{Minimum effective sample size for all parameters}

\item{\code{psrf_tol}}{upper threshold for Gelman-Rubin potential scale reduction factors (if diag is "gelman")}

\item{\code{max_attempts}}{Maximum number of retries for assessing convergence and generating enough effective samples.
Can be set to Inf to run indefinitely.}

\item{\code{nchains_diag}}{number of independent chains to run (including this one) to use
convergence diagnostics that require multiple chains.}

\item{\code{iterations}}{Number of MCMC iterations (not including burnin or thinning)}

\item{\code{burnin}}{MCMC burnin, the number of MCMC iterations to ignore at the beginning of each obj}

\item{\code{thin}}{MCMC thinning interval, number of iterations to skip in between iterations to return.}

\item{\code{verbose}}{Number of times to print out progress during burnin and the actual MCMC.
If \code{verbose} > 0, then progress messages will also print during setup and cleanup.}

\item{\code{mphtol}}{Tolerance theshold for mid-parent heterosis (on a log base e scale)}

\item{\code{returns}}{Character vector naming the variables whose MCMC samples
you want to return}

\item{\code{updates}}{Character vector naming the variables to calculate/update
during the MCMC.}

\item{\code{samples_return}}{Indices of RNA-seq samples/libraries whose parameter samples you want to return.
Applies to all library-specific parameters except for the epsilons.}

\item{\code{features_return}}{Indices of features/genes whose parameter samples you want to return.
Applies to all gene-specific parameters except for the epsilons.}

\item{\code{samples_return_eps}}{Indices of RNA-seq samples/libraries n for which epsilon_{n, g} is updated/returned.}

\item{\code{features_return_eps}}{Indices of features/genes g for which epsilon_{n, g} is updated/returned.}

\item{\code{phiPrior}}{Name of the family of priors on the phi_g's after integrating out the xi_phi's.
Can be any value returned by \code{alternate_priors()}. All other values will default to the normal prior.}

\item{\code{alpPrior}}{Name of the family of priors on the alp_g's after integrating out the xi_alp's.
Can be any value returned by \code{alternate_priors()}. All other values will default to the normal prior.}

\item{\code{delPrior}}{Name of the family of priors on the del_g's after integrating out the xi_del's.
Can be any value returned by \code{alternate_priors()}. All other values will default to the normal prior.}
}}

