% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configs.R
\name{Configs}
\alias{Configs}
\title{Constructor for class \code{Configs}}
\usage{
Configs(obj = NULL, ...)
}
\arguments{
\item{obj}{a \code{Chain} or \code{list} object to get slots from.}

\item{...}{optional slot values.}
}
\value{
a \code{Configs} object
}
\description{
Create a \code{Configs} from a \code{Chain} object or by
setting individual slots.
}
\seealso{
\code{help("Configs-class")}
}

