% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameters.R
\name{constants}
\alias{constants}
\title{Function \code{contstants}}
\usage{
constants()
}
\value{
Character vector of parameter names.
}
\description{
Prints out initialization constants of the model. See the methodology vignette for details.
}

