% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_estimates.R
\name{contrast_estimates}
\alias{contrast_estimates}
\title{Function \code{contrast_estimates}}
\usage{
contrast_estimates(obj, level = 0.95)
}
\arguments{
\item{obj}{a \code{Chain} object or list of \code{Chain} objects generated by \code{fbseq()}.}

\item{level}{level of the credible intervals from 0 to 1}
}
\value{
a data frame of marginal posterior parameter estimates: posterior means, poseterior standard deviations, and approximate credible intervals.
Parameters not updeted in the MCMC are excluded.
}
\description{
Extracts estimated posterior means, standard deviations and credible intervals
of linear combinations of beta parameters.
}

