% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DIC.R
\name{DIC}
\alias{DIC}
\title{Function \code{DIC}}
\usage{
DIC(chain_list)
}
\arguments{
\item{chain_list}{a \code{Chain} object or list of \code{Chain} objects generated by \code{fbseq()}}
}
\value{
DIC
}
\description{
Get the deviance information criterion (DIC) of a chain or list of chains, along with the
effective and actual numbers of parameters.
}

