% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_starts.R
\name{disperse_starts}
\alias{disperse_starts}
\title{Function \code{disperse_starts}}
\usage{
disperse_starts(chain)
}
\arguments{
\item{chain}{\code{Chain} object that has already been run with \code{run_mcmc()}.}
}
\value{
a \code{Chain} object with dispersed MCMC starting values.
}
\description{
Uses the hyperparameter samples from a previous chain to 
generate another \code{Chain} object with dispersed starting values.
Multiple \code{Chain} objects produced by \code{disperse_startes()} can
be run and Gelman potential scale reduction factors on the collective output
will make sense as convergence diagnostics
}

