% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_starts.R
\name{dispersed_set}
\alias{dispersed_set}
\title{Function \code{dispersed_set}}
\usage{
dispersed_set(chain, parm, lower = NA, upper = NA)
}
\arguments{
\item{chain}{\code{Chain} object that has already been run with \code{run_mcmc()}.}

\item{parm}{character string specifying the parameter set to work on.}

\item{lower}{lower bound for nonnegative parameters.}

\item{upper}{upper bound for nonnegative parameters.}
}
\value{
a vector with a dispersed starting values (relative to the estimated posterior so far)
for a given set of parameters.
}
\description{
Produces a dispersed starting values (relative to the estimated posterior so far)
for a given set of parameters.
}

