% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/effective_sample_size.R
\name{effective_sample_size}
\alias{effective_sample_size}
\title{Function \code{effective_sample_size}}
\usage{
effective_sample_size(chain)
}
\arguments{
\item{chain}{a \code{Chain} object}
}
\value{
a \code{Chain} object
}
\description{
Reruns an already-run \code{Chain} object until
at least \code{chain@iterations} effective samples are obtained for each returned hyperparameter.
}

