% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/effectiveSampleSize.R
\name{effectiveSampleSize}
\alias{effectiveSampleSize}
\title{Function \code{effectiveSampleSize}}
\usage{
effectiveSampleSize(chain)
}
\arguments{
\item{chain}{a \code{Chain} object}
}
\value{
a \code{Chain} object
}
\description{
Reruns an already-run \code{Chain} object until
at least \code{chain@iterations} effective samples are obtained for each returned hyperparameter.
}

