% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates}
\alias{estimates}
\title{Function \code{estimates}}
\usage{
estimates(chain, level = 0.95)
}
\arguments{
\item{chain}{a \code{Chain} object}

\item{level}{level of the credible intervals from 0 to 1}
}
\value{
a data frame of estimates: posterior means and approximate credible intervals.
Parameters not updeted in the MCMC are excluded.
}
\description{
Extracts posterior means and estimated credible intervals
from a \code{Chain} object.
}

