% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_generated_data}
\alias{example_generated_data}
\title{example_generated_data}
\description{
Example generated data generated with \code{generate_data()}.
}
\author{
Will Landau \email{will.landau@gmail.com}
}

