% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chain_constructor.R
\name{fill_easy_gaps}
\alias{fill_easy_gaps}
\title{Function \code{fill_easy_gaps}}
\usage{
fill_easy_gaps(chain, counts, group)
}
\arguments{
\item{chain}{a \code{Chain} object}

\item{counts}{Matrix of RNA-seq read counts.}

\item{group}{Experimental design. A vector of integers,
one for each RNA-seq sample/library, denoting the genetic
variety of that sample. You must use 1 for parent 1, 2 for parent 2,
and 3 for the hybrid.}
}
\value{
a \code{Chain} object
}
\description{
Set required slots in a \code{Chain} object.
Used internally.
}

