% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten}
\alias{flatten}
\title{Function \code{flatten}}
\usage{
flatten(obj)
}
\arguments{
\item{obj}{The argument, either a \code{Chain} or a \code{Starts} object.}
}
\value{
one of the following, depending on the class of the argument object.
\describe{
  \item{\code{Chain} argument}{a data frame parameter libraries.}
  \item{\code{Starts} argument}{a named numeric vector of starting values. If
\code{epsilon} is set in \code{obj}, then one of \code{rho}, \code{phi}, \code{alp},
or \code{gamma} must be set too.}
}
}
\description{
Flattens an object into some kind of useful array.
}

