% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten_chain}
\alias{flatten_chain}
\title{Function \code{flatten_chain}}
\usage{
flatten_chain(chain)
}
\arguments{
\item{chain}{a \code{Chain} object.}
}
\value{
an array of MCMC parameter libraries.
}
\description{
Flattens a \code{Chain} object.
}

