% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten_post}
\alias{flatten_post}
\title{Function \code{flatten_post}}
\usage{
flatten_post(chain, square = F, updated_only = T)
}
\arguments{
\item{chain}{a \code{Chain} object.}

\item{square}{TRUE/FALSE value. If TRUE, return the flattened posterior mean squares
instead of posterior means}

\item{updated_only}{TRUE/FALSE. If TRUE, return the posterior quantities only for the 
parameters that were updated in the MCMC.}
}
\value{
an array of MCMC parameter libraries.
}
\description{
Flattens posterior means or posterior means of squares
}

