% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten_starts}
\alias{flatten_starts}
\title{Function \code{flatten_starts}}
\usage{
flatten_starts(starts)
}
\arguments{
\item{starts}{a \code{Starts} object.}
}
\value{
an array of MCMC parameter libraries.
}
\description{
Flattens a \code{Starts} object.
}

