% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gelman_attempt.R
\name{gelman_attempt}
\alias{gelman_attempt}
\title{Function \code{gelman_attempt}}
\usage{
gelman_attempt(chain_list, pattern)
}
\arguments{
\item{chain_list}{a list of \code{Chain} objects}

\item{pattern}{character string: grep pattern for the parameters of interest}
}
\value{
a list of \code{Chain} objects
}
\description{
Run 4 chains together and compute and report Gelman factors.
}

