% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Function \code{generate_data}}
\usage{
generate_data(libraries = 12, genes = 35000, design = cbind(rep(1,
  libraries), rep(c(1, -1, 1), each = floor(libraries/3)), rep(c(-1, 1, 1), each
  = floor(libraries/3))), starts = Starts(nuGamma = 5, nuRho = 5, omega = c(1,
  0.5, 0.5), tauGamma = 1, tauRho = 0.1, theta = c(3, 0, 0)))
}
\arguments{
\item{libraries}{number of libraries/libraries in the data}

\item{genes}{number of genes/genes in the data}

\item{starts}{A \code{Starts} object containing the necessary initialization constants
and hyperparameters.}

\item{group}{Experimental design. A vector of integers,
one for each RNA-seq sample/library, denoting the genetic
variety of that sample. You must use 1 for parent 1, 2 for the hybrid,
and 3 for parent 2.}

\item{priors}{Name of the family of priors on the betas. Must be length 1 or ncol(design).
Can be among "Laplace", "t", or "horseshoe". All other values will default to the normal prior.}
}
\value{
a list
\describe{
  \item{\code{counts}}{The generated RNA-seq counts matrix.}
  \item{\code{design}}{The design matrix for gene-specific effects.}
  \item{\code{truth}}{A \code{Starts} object containing the values used to simulate the data}
}
}
\description{
Simulates heterosis data from prespecified hyperparameters.
}

