% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_data_from_model.R
\name{generate_data_from_model}
\alias{generate_data_from_model}
\title{Function \code{generate_data_from_model}}
\usage{
generate_data_from_model(genes, design, truth, priors = "Laplace")
}
\arguments{
\item{genes}{number of genes/features/response variables in the data}

\item{design}{Design matrix, with \code{N} rows and \code{L} columns.
Same format as in \code{edgeR::glmFit}, with RNA-seq libraries/samples as rows
and each \code{beta_{l, .}} as a column.}

\item{truth}{a \code{Starts} object containing the necessary initialization constants
and hyperparameters. All hyperparameters must be supplied.}

\item{priors}{character string, either a member of \code{special_beta_priors()} or
another string for normal priors on the betas.}
}
\value{
a list: a \code{Scenario} object with the scenario and a \code{Starts} object
with the true parameters that the count dataset in the \code{Scenario} object
was simulated from.
}
\description{
Simulates count data fromt the Laplace-prior version of the model
given hyperparameters. All hyperparameters must be given in the \code{truth} slot
of the \code{Scenario} object argument.
}

