% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_starts.R
\name{get_nonzeros}
\alias{get_nonzeros}
\title{Function \code{get_nonzeros}}
\usage{
get_nonzeros(x)
}
\arguments{
\item{x}{argument}
}
\value{
x with no zeros
}
\description{
Helper function for \code{generate_starts()}.
}

