% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuda_usage.R
\name{getDevice}
\alias{getDevice}
\title{Function \code{getDevice}}
\usage{
getDevice()
}
\value{
Integer index of the current CUDA-capable GPU, which is >= 0 and 
< number of devices.
}
\description{
Get the index of the current CUDA-capable GPU.
}

