% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heterosis.R
\name{heterosis}
\alias{heterosis}
\title{Function \code{heterosis}}
\usage{
heterosis(chain)
}
\arguments{
\item{chain}{object of type \code{Chain}. See the package vignette for details.}
}
\value{
a \code{Chain} object with updated parameter values. You can feed
this object back into another call to \code{run_mcmc()} to continue the chain
from where you left off.
}
\description{
Top-level function. Runs an MCMC with the runtime
parameters specified in the \code{Chain} object
}

