% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameters.R
\name{hyperparameters}
\alias{hyperparameters}
\title{Function \code{hyperparameters}}
\usage{
hyperparameters()
}
\value{
Character vector of hyperparameter names.
}
\description{
Same as \code{parameters()}, except that only the names
of hyperparameters are printed out. Type \code{?parameters} for details.
}

