% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inference.R
\docType{class}
\name{Inference-class}
\alias{Inference-class}
\title{Class \code{Inference}}
\description{
A collection of MCMC control parameters.
}
\section{Slots}{

\describe{
\item{\code{conjunctions}}{List of numeric vectors of length J. Each numeric vector is a collection of
indices c_1, ..., c_J such that the posterior probability,
P(contrasts[[c_1]] * beta > value[c_1], ..., contrasts[[c_J]] * beta > value[c_J] | data),
is estimated by averging over Monte Carlo samples. This allows the user to "test"
multiple contrasts together. See the vignette for a more thorough explanation.}

\item{\code{contrasts}}{list of numeric vectors, each length \code{L}. Each vector
is a contrast of the \code{beta_{l, .}}'s of interest.}

\item{\code{design}}{Design matrix, with \code{N} rows and \code{L} columns.
Same format as in \code{edgeR::glmFit}, with RNA-seq libraries/samples as rows
and each \code{beta_{l, .}} as a column.}

\item{\code{probs}}{matrix with G rows and J columns. These are the posterior probabilities
of each conjunction above for each gene, estimated by averaging over Monte Carlo
samples.}

\item{\code{values}}{Numeric vector of length \code{length(contrasts)}. Used int tests.}
}}

