% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inference.R
\name{Inference}
\alias{Inference}
\title{Constructor for class \code{Inference}}
\usage{
Inference(obj = NULL, ...)
}
\arguments{
\item{obj}{a \code{Chain} or \code{list} object to get slots from.}

\item{...}{optional slot values.}
}
\value{
a \code{Inference} object
}
\description{
Constructor for class \code{Inference}
}
\details{
Precedence will be given to the \code{Chain} or \code{list} object over \code{...}.
}

