% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcmc_samples.R
\name{mcmc_samples}
\alias{mcmc_samples}
\title{Function \code{mcmc_samples}}
\usage{
mcmc_samples(obj)
}
\arguments{
\item{obj}{a \code{Chain} object or list of \code{Chain} objects generated by \code{fbseq()}.}
}
\value{
a data frame of MCMC samples
}
\description{
Get MCMC samples from a \code{Chain} object
or a list of \code{Chain} objects generated by \code{fbseq()}.
}

