% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_starts.R
\name{nu_tau}
\alias{nu_tau}
\title{Function \code{nu_tau}}
\usage{
nu_tau(x)
}
\arguments{
\item{x}{argument}
}
\value{
list of nu and tau estimates
}
\description{
Helper function for \code{simple_starts()}.
}

