% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{Function \code{parameters}}
\usage{
parameters()
}
\value{
Character vector of parameter names.
}
\description{
Prints out the full collection of variable names. See the methodology vignette for details.
}

