% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paschold}
\alias{paschold}
\title{paschold}
\description{
a \code{Scenario} object guiding the analysis of the Paschold et al (2012) dataset 
(http://www.ncbi.nlm.nih.gov/pubmed/23086286).
}
\author{
Paschold et al. (http://www.ncbi.nlm.nih.gov/pubmed/23086286)
}

