% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paschold_group}
\alias{paschold_group}
\title{paschold_group}
\description{
A vector of integers denoting the experimental design.
There is one integer for each RNA-seq sample/library, each denoting
 the genetic variety of that sample: 1 for parent 1, 2 for the hybrid,
and 3 for parent 2.
}
\author{
Will Landau \email{will.landau@gmail.com}
}

