% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/probs.R
\name{probs}
\alias{probs}
\title{Function \code{probs}}
\usage{
probs(chain)
}
\arguments{
\item{chain}{a \code{Chain} object}
}
\value{
a data frame of estimated posterior probabilities, with rows corresponding to genes and columns
corresponding to the propositions in \code{Scenario(chain)@propositions}.
}
\description{
Extracts estimated posterior probs from a \code{Chain} object.
}

