% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/psrf.R
\name{psrf}
\alias{psrf}
\title{Function \code{psrf}}
\usage{
psrf(chain, important = T, sort = T, threshold = 1.1)
}
\arguments{
\item{chain}{a \code{Chain} object}

\item{important}{If \code{TRUE}, only the important psrf will be returned. The
important psrf correspond to the hyperparameters (thetas, sigmas, nu, and tau)
plus the betas used in estimating the posterior probabilities specified in \code{Scenario(chain)}.}

\item{sort}{sort the psrf in decreasing order}

\item{threshold}{If only the psrf above \code{threshold} will be returned. To return all psrf,
set \code{threshold} to 0.}
}
\value{
a vector of important potential scale reduction factors
}
\description{
Get the (important) Gelman-Rubin potential scale reduction factors (psrf) from a \code{Chain}
object.
}

