% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrf_all.R
\name{psrf_all}
\alias{psrf_all}
\title{Function \code{psrf_all}}
\usage{
psrf_all(chain_list)
}
\arguments{
\item{chain_list}{list of \code{Chain} objects}
}
\value{
Gelman potential scale reduction factors on a list of parallel independent chains.
}
\description{
calculate all Gelman potential scale reduction factors on a list of parallel independent chains.
}

