% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_gelman_mcmc.R
\name{run_gelman_mcmc}
\alias{run_gelman_mcmc}
\title{Function \code{run_gelman_mcmc}}
\usage{
run_gelman_mcmc(chain)
}
\arguments{
\item{chain}{a \code{Chain} object}
}
\value{
a \code{Chain} object
}
\description{
Accepts RNA-seq count data and group structure as arguments and
returns a set of heterosis probabilities.
}
\details{
Multiple MCMC chain is run until Gelman potential scale reduction factors
show no evidence of a lack of convergence.
}

