% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampler_options.R
\name{sampler_options}
\alias{sampler_options}
\title{Function \code{sampler_options}}
\usage{
sampler_options()
}
\value{
vector of options for slot of the \code{Samplers} object
}
\description{
Returns the vector of options for slot of the \code{Samplers} object
}

