% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samplers.R
\docType{class}
\name{Samplers-class}
\alias{Samplers-class}
\title{Class \code{Samplers}}
\description{
specification of the MCMC black box sampler of each parameter
}
\section{Slots}{

\describe{
\item{\code{beta}}{sampler specification (character vector of length 1, element of samplers())}

\item{\code{epsilon}}{sampler specification (character vector of length 1, element of samplers())}

\item{\code{gamma}}{sampler specification (character vector of length 1, element of samplers())}

\item{\code{nu}}{sampler specification (character vector of length 1, element of samplers())}

\item{\code{sigmaSquared}}{sampler specification (character vector of length 1, element of samplers())}

\item{\code{tau}}{sampler specification (character vector of length 1, element of samplers())}

\item{\code{theta}}{sampler specification (character vector of length 1, element of samplers())}

\item{\code{xi}}{sampler specification (character vector of length 1, element of samplers())}
}}

