% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{samplers}
\alias{samplers}
\title{Function \code{samplers}}
\usage{
samplers()
}
\value{
vector of options for slot of the \code{Samplers} object
}
\description{
Returns the vector of options for slot of the \code{Samplers} object. See the tutorial vignette for details.
}

