% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\docType{class}
\name{Scenario-class}
\alias{Scenario-class}
\title{Class \code{Scenario}}
\description{
A collection of MCMC control parameters.
}
\section{Slots}{

\describe{
\item{\code{bounds}}{Numeric vector of length \code{length(contrasts)}. Used to formulate 
logical propositions involving contrasts.}

\item{\code{contrasts}}{list of numeric vectors, each length \code{L}. Each vector
is a contrast of the \code{beta_{l, .}}'s of interest.}

\item{\code{counts}}{RNA-seq count data matrix with G rows (for genes) and N columns (for libraries/samples).}

\item{\code{design}}{Design matrix, with \code{N} rows and \code{L} columns.
Same format as in \code{edgeR::glmFit}, with RNA-seq libraries/samples as rows
and each \code{beta_{l, .}} as a column.}

\item{\code{propositions}}{List of numeric vectors of length \code{P}. Each numeric vector \code{v} denotes a logical
proposition that involves the \code{contrasts} list and \code{bounds} vector. Each element of \code{v}
is the index of a contrast in \code{contrasts}. For each gene, the MCMC will estimate the posterior
probability that each contrast (of the \code{beta} parameters) denoted in \code{v} is greater than 
its corresponding value in \code{bound}. See the tutorial vignette for more details}

\item{\code{supplement}}{a list containing supplementary information about the scenario: 
for example, how the data were simulated, if applicable}
}}

