% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{Scenario}
\alias{Scenario}
\title{Constructor for class \code{Scenario}}
\usage{
Scenario(obj = NULL, ...)
}
\arguments{
\item{obj}{a \code{Chain} or \code{list} object to get slots from.}

\item{...}{optional slot bounds.}
}
\value{
a \code{Scenario} object
}
\description{
Create a \code{Scenario} from a \code{Chain} object or by
setting individual slots.
}
\seealso{
\code{help("Scenario-class")}
}

