% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cuda_usage.R
\name{setDevice}
\alias{setDevice}
\title{Function \code{setDevice}}
\usage{
setDevice(device)
}
\arguments{
\item{device}{Integer index of a CUDA-capable GPU. Must be >= 0 < number of GPUs.}
}
\value{
Integer index of the newly-set CUDA-capable GPU, which is >= 0 and
< number of devices.
}
\description{
Choose a CUDA-capable GPU to use for the MCMC.
}

