% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_starts.R
\name{simple_starts}
\alias{simple_starts}
\title{Function \code{simple_starts}}
\usage{
simple_starts(chain, counts, group)
}
\arguments{
\item{chain}{\code{Chain} object whose starting values to fill.}

\item{counts}{A data frame/matrix of RNA-seq read counts or list of slots. If a list,
then the function will return a \code{Chain} object with those slots.}

\item{group}{Experimental design. A vector of integers,
one for each RNA-seq sample/library, denoting the genetic
variety of that sample. You must use 1 for parent 1, 2 for parent 2,
and 3 for the hybrid.}
}
\value{
a \code{Chain} object with a full set of MCMC starting values.
}
\description{
Generates MCMC starting values using a \code{Starts} object and an
\code{Internals} object. The \code{configure(...)} function must be called first.
}
\details{
Simple nonstochastic starting values are calculated by quick arithmetic.
}

