% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_mcmc.R
\name{single_mcmc}
\alias{single_mcmc}
\title{Function \code{single_mcmc}}
\usage{
single_mcmc(chain, backend = "CUDA", threads = 1)
}
\arguments{
\item{chain}{object of type \code{Chain}. This \code{chain} argument could be
a newly created \code{Chain} object from \code{Chain(...)}. Alternatively,
if \code{chain} is the output from a previous call to \code{run_mcmc(...)},
then the function will continue the MCMC from where it left off.}

\item{backend}{defaults to "CUDA". Other options include "serial", which
does not use any parallel computing, and "OpenMP", which uses "OpenMP" 
threads to parallelize within Markov chains.}

\item{threads}{number of threads for the OpenMP backend.}
}
\value{
a \code{Chain} object with updated parameter values. You can feed
this object back into another call to \code{single_mcmc()} to continue the chain
from where you left off.
}
\description{
Runs a single MCMC chain for a fixed number of iterations.
}

