% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallchain.R
\name{smallchain}
\alias{smallchain}
\title{Function \code{smallchain}}
\usage{
smallchain(size = 25)
}
\arguments{
\item{size}{Number of genes. Also the number of iterations and burnin.}
}
\value{
A small \code{Chain} object
}
\description{
Make a small \code{Chain} object for debugging and testing.
}

