% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/special_beta_priors.R
\name{special_beta_priors}
\alias{special_beta_priors}
\title{Function \code{special_beta_priors}}
\usage{
special_beta_priors()
}
\value{
string vector of names of special_beta priors
}
\description{
Function \code{special_beta_priors}
}
\details{
special priors on the beta_{l, g} parameters (using the xi's as aux variables)
}

