% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/starts.R
\docType{class}
\name{Starts-class}
\alias{Starts-class}
\title{Class \code{Starts}}
\description{
A way for users to specify model parameter
starting values. Use only when creating new \code{obj}
objects from scratch.
}
\section{Slots}{

\describe{
\item{\code{dRho}}{initialization constant}

\item{\code{dGam}}{initialization constant}

\item{\code{aRho}}{initialization constant}

\item{\code{aGam}}{initialization constant}

\item{\code{bRho}}{initialization constant}

\item{\code{bGam}}{initialization constant}

\item{\code{cPhi}}{initialization constant}

\item{\code{cAlp}}{initialization constant}

\item{\code{cDel}}{initialization constant}

\item{\code{sPhi}}{initialization constant}

\item{\code{sAlp}}{initialization constant}

\item{\code{sDel}}{initialization constant}

\item{\code{kPhi}}{initialization constant}

\item{\code{kAlp}}{initialization constant}

\item{\code{kDel}}{initialization constant}

\item{\code{rPhi}}{initialization constant}

\item{\code{rAlp}}{initialization constant}

\item{\code{rDel}}{initialization constant}

\item{\code{nuRho}}{model parameter}

\item{\code{nuGam}}{model parameter}

\item{\code{tauRho}}{model parameter}

\item{\code{tauGam}}{model parameter}

\item{\code{thePhi}}{model parameter}

\item{\code{theAlp}}{model parameter}

\item{\code{theDel}}{model parameter}

\item{\code{sigPhi}}{model parameter}

\item{\code{sigAlp}}{model parameter}

\item{\code{sigDel}}{model parameter}

\item{\code{phi}}{model parameter}

\item{\code{alp}}{model parameter}

\item{\code{del}}{model parameter}

\item{\code{rho}}{model parameter}

\item{\code{gam}}{model parameter}

\item{\code{xiPhi}}{model parameter}

\item{\code{xiAlp}}{model parameter}

\item{\code{xiDel}}{model parameter}

\item{\code{eps}}{model parameter}
}}

