% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chain.R
\docType{class}
\name{Chain-class}
\alias{Chain-class}
\title{Class \code{Chain}}
\description{
the main storage object for a heterosis MCMC chain.
Create a new one with a call to \code{Chain()} and run MCMC
by feeding one into \code{single_chain()}.
}
\section{Slots}{

\describe{
\item{\code{diag}}{convergence diagnostic to use. Can be "gelman" or "none".}

\item{\code{ess}}{Minimum effective sample size for all parameters}

\item{\code{max_attempts}}{Maximum number of retries for assessing convergence and generating enough effective samples.
Can be set to Inf to run indefinitely.}

\item{\code{nchains_diag}}{number of independent chains to run (including this one) to use
convergence diagnostics that require multiple chains.}

\item{\code{psrf_tol}}{upper threshold for Gelman-Rubin potential scale reduction factors (if diag is "gelman")}

\item{\code{burnin}}{MCMC burnin, the number of MCMC iterations to ignore at the beginning of each obj}

\item{\code{genes_return}}{Indices of genes whose parameter samples you want to return.
Applies to all gene-specific parameters except for the epsilons.}

\item{\code{genes_return_epsilon}}{Indices of genes g for which epsilon_{n, g} is updated/returned.}

\item{\code{iterations}}{Number of MCMC iterations (not including burnin or thinning)}

\item{\code{libraries_return}}{Indices of RNA-seq libraries whose parameter samples you want to return.}

\item{\code{libraries_return_epsilon}}{Indices of RNA-seq libraries n for which epsilon_{n, g} is updated/returned.
Applies to all library-specific parameters except for the epsilons.}

\item{\code{parameter_sets_return}}{Character vector naming the variables whose MCMC samples
you want to return}

\item{\code{parameter_sets_update}}{Character vector naming the variables to calculate/update
during the MCMC.}

\item{\code{priors}}{Names of the family of priors on the betas after integrating out the xi's.
Can be any value returned by \code{alternate_priors()}. All other values will default to the normal prior.}

\item{\code{thin}}{MCMC thinning interval, number of iterations to skip in between iterations to return.}

\item{\code{verbose}}{Number of times to print out progress during burnin and the actual MCMC.
If \code{verbose} > 0, then progress messages will also print during setup and cleanup.}

\item{\code{counts}}{RNA-seq count data, flattened from a matrix}

\item{\code{countSums_g}}{gene-specific count sums}

\item{\code{countSums_n}}{library-specific count sums}

\item{\code{design}}{Gene-specific design, flattened from the design matrix. Must contain only 0's, 1's, and -1's.
Original matrix must have rows corresponding to colums/libraries in RNA-seq data and colums corresponding to
gene-specific variables.}

\item{\code{G}}{number of genes}

\item{\code{Greturn}}{number of genes to return gene-specific MCMC parameter samples for (except the epsilons)}

\item{\code{GreturnEpsilon}}{number of genes to return gene-specific MCMC epsilon parameter samples}

\item{\code{L}}{number of columns in the original design matrix}

\item{\code{N}}{number of libraries}

\item{\code{Nreturn}}{number of libraries to return library-specific MCMC parameter samples for (except the epsilons)}

\item{\code{NreturnEpsilon}}{number of libraries to return library-specific MCMC epsilon parameter samples}

\item{\code{psrf}}{Gelman-Rubin potential scale reduction factors for the sampled parameters (even if the
actual MCMC parameter samples are not returned)}

\item{\code{seeds}}{vector of N*G random number generator seeds}

\item{\code{aRho}}{initialization constant}

\item{\code{aGamma}}{initialization constant}

\item{\code{bRho}}{initialization constant}

\item{\code{bGamma}}{initialization constant}

\item{\code{c}}{initialization constants}

\item{\code{dRho}}{initialization constant}

\item{\code{dGamma}}{initialization constant}

\item{\code{k}}{initialization constants}

\item{\code{r}}{initialization constants}

\item{\code{s}}{initialization constants}

\item{\code{beta}}{MCMC parameter samples}

\item{\code{epsilon}}{MCMC parameter samples}

\item{\code{gamma}}{MCMC parameter samples}

\item{\code{nuGamma}}{MCMC parameter samples}

\item{\code{nuRho}}{MCMC parameter samples}

\item{\code{omega}}{MCMC parameter samples}

\item{\code{rho}}{MCMC parameter samples}

\item{\code{tauGamma}}{MCMC parameter samples}

\item{\code{tauRho}}{MCMC parameter samples}

\item{\code{theta}}{MCMC parameter samples}

\item{\code{xi}}{MCMC parameter samples}

\item{\code{betaStart}}{MCMC starting values}

\item{\code{epsilonStart}}{MCMC starting values}

\item{\code{gammaStart}}{MCMC starting values}

\item{\code{nuRhoStart}}{MCMC starting values}

\item{\code{nuGammaStart}}{MCMC starting values}

\item{\code{omegaStart}}{MCMC starting values}

\item{\code{rhoStart}}{MCMC starting values}

\item{\code{tauRhoStart}}{MCMC starting values}

\item{\code{tauGammaStart}}{MCMC starting values}

\item{\code{thetaStart}}{MCMC starting values}

\item{\code{xiStart}}{MCMC starting values}

\item{\code{betaPostMean}}{estimated posterior means}

\item{\code{epsilonPostMean}}{estimated posterior means}

\item{\code{gammaPostMean}}{estimated posterior means}

\item{\code{nuRhoPostMean}}{estimated posterior mean}

\item{\code{nuGammaPostMean}}{estimated posterior mean}

\item{\code{omegaPostMean}}{estimated posterior means}

\item{\code{rhoPostMean}}{estimated posterior means}

\item{\code{tauRhoPostMean}}{estimated posterior mean}

\item{\code{tauGammaPostMean}}{estimated posterior mean}

\item{\code{thetaPostMean}}{estimated posterior means}

\item{\code{xiPostMean}}{estimated posterior means}

\item{\code{betaPostMeanSquare}}{estimated posterior means of the squares of parameters}

\item{\code{epsilonPostMeanSquare}}{estimated posterior means of the squares of parameters}

\item{\code{gammaPostMeanSquare}}{estimated posterior means of the squares of parameters}

\item{\code{nuRhoPostMeanSquare}}{estimated posterior mean of the square of the parameter}

\item{\code{nuGammaPostMeanSquare}}{estimated posterior mean of the square of the parameter}

\item{\code{omegaPostMeanSquare}}{estimated posterior means of the squares of parameters}

\item{\code{rhoPostMeanSquare}}{estimated posterior means of the squares of parameters}

\item{\code{tauRhoPostMeanSquare}}{estimated posterior mean of the square of the parameter}

\item{\code{tauGammaPostMeanSquare}}{estimated posterior mean of the square of the parameter}

\item{\code{thetaPostMeanSquare}}{estimated posterior means of the squares of parameters}

\item{\code{xiPostMeanSquare}}{posterior means of the squares of parameters}
}}

