% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chain.R
\docType{class}
\name{Chain-class}
\alias{Chain-class}
\title{Class \code{Chain}}
\description{
the main storage object for a heterosis MCMC chain.
Create a new one with a call to \code{Chain()} and run MCMC
by feeding one into \code{single_chain()}.
}
\section{Slots}{

\describe{
\item{\code{bound_names}}{names of \code{bounds} slot in the original \code{Scenario} object}

\item{\code{contrast_names}}{names of \code{contrasts} slot in the \code{Scenario} object}

\item{\code{gene_names}}{names of genes, taken from the row names of the count data matrix}

\item{\code{library_names}}{names of the libraries/samples, taken from the column names of the count data matrix}

\item{\code{proposition_names}}{names of \code{propositions} slot in the \code{Scenario} object}

\item{\code{bounds}}{values to compare contrasts to. The comparison is to see if the contrast is greater than
its corresponding element in \code{bounds} (from the \code{Scenario} object)}

\item{\code{contrasts}}{contrasts from the \code{Scenario} object.}

\item{\code{counts}}{RNA-seq count data, flattened from a matrix}

\item{\code{design}}{Design, flattened from the design matrix.
Original matrix must have rows corresponding to colums/libraries in RNA-seq data and colums corresponding to
sets of gene-specific variables.}

\item{\code{propositions}}{propositions of inequalities involving contrasts from the \code{Scenario} object.}

\item{\code{supplement}}{a list containing supplementary information about the scenario:
for example, how the data were simulated, if applicable}

\item{\code{burnin}}{MCMC burnin, the number of MCMC iterations to ignore at the beginning of each obj}

\item{\code{effects_update_beta}}{bounds of l for which to update the beta_{l, g} parameters.}

\item{\code{genes_return}}{Indices of genes whose parameter samples you want to return.
Applies to all gene-specific parameters except for the epsilons.}

\item{\code{genes_return_epsilon}}{Indices of genes g for which epsilon_{n, g} is updated/returned.}

\item{\code{iterations}}{Number of MCMC iterations after burnin for which selected parameter samples are kept.
Total MCMC iterations = burnin + thin * "iterations", and the whole "thin * iterations" portion
is used to calculate posterior means, mean squares, and probabilities.}

\item{\code{libraries_return}}{Indices of RNA-seq libraries whose parameter samples you want to return.}

\item{\code{libraries_return_epsilon}}{Indices of RNA-seq libraries n for which epsilon_{n, g} is updated/returned.
Applies to all library-specific parameters except for the epsilons.}

\item{\code{parameter_sets_return}}{Character vector naming the variables whose MCMC samples
you want to return}

\item{\code{parameter_sets_update}}{Character vector naming the variables to calculate/update
during the MCMC.}

\item{\code{priors}}{Names of the family of priors on the betas after integrating out the xi's.
Can be any value returned by \code{special_beta_priors()}. All other bounds will default to the normal prior.}

\item{\code{thin}}{MCMC thinning interval. \code{thin = 1} means parameter samples will be saved for every iterations
after burnin. \code{thin = 10} means parameter samples will be saved every 10th iteration after burnin.
Total MCMC iterations = burnin + thin * "iterations", and the whole "thin * iterations" portion
is used to calculate posterior means, mean squares, and probabilities.}

\item{\code{verbose}}{Number of times to print out progress during burnin and the actual MCMC.
If \code{verbose} > 0, then progress messages will also print during setup and cleanup.}

\item{\code{C}}{number of contrasts}

\item{\code{countSums_g}}{gene-specific count sums}

\item{\code{countSums_n}}{library-specific count sums}

\item{\code{designUnique}}{Matrix of unique nonzero elements of \code{design}. Vacent entries are 0.}

\item{\code{designUniqueN}}{for each column index \code{l}, number of unique nonzero elements of \code{design[, l]}.}

\item{\code{G}}{number of genes}

\item{\code{Greturn}}{number of genes to return gene-specific MCMC parameter samples for (except the epsilons)}

\item{\code{GreturnEpsilon}}{number of genes to return gene-specific MCMC epsilon parameter samples}

\item{\code{L}}{number of columns in the original design matrix}

\item{\code{Lupdate_beta}}{number of bounds of l for which to update the beta_{l, g} parameters.}

\item{\code{N}}{number of libraries}

\item{\code{Nreturn}}{number of libraries to return library-specific MCMC parameter samples for (except the epsilons)}

\item{\code{NreturnEpsilon}}{number of libraries to return library-specific MCMC epsilon parameter samples}

\item{\code{P}}{number of propositions involving contrasts}

\item{\code{probs}}{estimated posterior probabilities of propositions in the \code{Scenario} object.}

\item{\code{seeds}}{vector of N*G random number generator seeds}

\item{\code{a}}{initialization constant}

\item{\code{b}}{initialization constant}

\item{\code{c}}{initialization constants}

\item{\code{d}}{initialization constant}

\item{\code{h}}{initialization constants}

\item{\code{k}}{initialization constants}

\item{\code{q}}{initialization constants}

\item{\code{r}}{initialization constants}

\item{\code{s}}{initialization constants}

\item{\code{beta}}{MCMC parameter samples}

\item{\code{epsilon}}{MCMC parameter samples}

\item{\code{gamma}}{MCMC parameter samples}

\item{\code{nu}}{MCMC parameter samples}

\item{\code{sigmaSquared}}{MCMC parameter samples}

\item{\code{tau}}{MCMC parameter samples}

\item{\code{theta}}{MCMC parameter samples}

\item{\code{xi}}{MCMC parameter samples}

\item{\code{betaStart}}{MCMC starting bounds}

\item{\code{epsilonStart}}{MCMC starting bounds}

\item{\code{gammaStart}}{MCMC starting bounds}

\item{\code{nuStart}}{MCMC starting bounds}

\item{\code{sigmaSquaredStart}}{MCMC starting bounds}

\item{\code{tauStart}}{MCMC starting bounds}

\item{\code{thetaStart}}{MCMC starting bounds}

\item{\code{xiStart}}{MCMC starting bounds}

\item{\code{betaPostMean}}{estimated posterior means}

\item{\code{epsilonPostMean}}{estimated posterior means}

\item{\code{gammaPostMean}}{estimated posterior means}

\item{\code{nuPostMean}}{estimated posterior mean}

\item{\code{sigmaSquaredPostMean}}{estimated posterior means}

\item{\code{tauPostMean}}{estimated posterior mean}

\item{\code{thetaPostMean}}{estimated posterior means}

\item{\code{xiPostMean}}{estimated posterior means}

\item{\code{betaPostMeanSquare}}{estimated posterior means of the squares of parameters}

\item{\code{epsilonPostMeanSquare}}{estimated posterior means of the squares of parameters}

\item{\code{gammaPostMeanSquare}}{estimated posterior means of the squares of parameters}

\item{\code{nuPostMeanSquare}}{estimated posterior mean of the square of the parameter}

\item{\code{sigmaSquaredPostMeanSquare}}{estimated posterior means of the squares of parameters}

\item{\code{tauPostMeanSquare}}{estimated posterior mean of the square of the parameter}

\item{\code{thetaPostMeanSquare}}{estimated posterior means of the squares of parameters}

\item{\code{xiPostMeanSquare}}{posterior means of the squares of parameters}
}}

