% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chain_constructor.R
\name{Chain}
\alias{Chain}
\title{Function \code{Chain}}
\usage{
Chain(data, design, configs = Configs(), starts = Starts())
}
\arguments{
\item{data}{A data frame/matrix of RNA-seq read counts or list of slots. If a list,
then the function will return a \code{Chain} object with those slots. Note: \code{data}
should only be a list within internal functions of the package. It is not recommended that
the user assign \code{data} to be a list.}

\item{design}{Gene-specific design matrix. Must contain only 0's, 1's, and -1's.
Must have rows corresponding to colums/libraries in RNA-seq data and colums corresponding to
gene-specific variables.}

\item{configs}{A \code{Configs} object of MCMC control parameters.}

\item{starts}{A \code{Starts} object of model parameter starting values.}
}
\value{
a \code{Chain} object ready to be passed to \code{single_chain()}
}
\description{
Construct a \code{Chain} object, the
one argument to \code{single_chain(...)}.
}
\details{
If \code{list} is set, all other arguments will be dropped.
}

