% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chain_constructor.R
\name{Chain}
\alias{Chain}
\title{Function \code{Chain}}
\usage{
Chain(data, group, configs = Configs(), starts = Starts())
}
\arguments{
\item{data}{A data frame/matrix of RNA-seq read counts or list of slots. If a list,
then the function will return a \code{Chain} object with those slots. Note: \code{data}
should only be a list within internal functions of the package. It is not recommended that
the user assign \code{data} to be a list.}

\item{group}{Experimental design. A vector of integers,
one for each RNA-seq sample/library, denoting the genetic
variety of that sample. You must use 1 for parent 1, 2 for parent 2,
and 3 for the hybrid.}

\item{configs}{A \code{Configs} object of MCMC control parameters.}

\item{starts}{A \code{Starts} object of model parameter starting values.}
}
\value{
a \code{Chain} object ready to be passed to \code{single_chain()}
}
\description{
Construct a \code{Chain} object, the
one argument to \code{single_chain(...)}.
}
\details{
If \code{list} is set, all other arguments will be dropped.
}

