% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chain_constructor.R
\name{Chain}
\alias{Chain}
\title{Function \code{Chain}}
\usage{
Chain(scenario, configs = Configs(), starts = Starts(), slots = NULL)
}
\arguments{
\item{scenario}{a \code{Scenario} object with count data, the design matrix, etc.}

\item{configs}{A \code{Configs} object of MCMC control parameters.}

\item{starts}{A \code{Starts} object of model parameter starting values.}

\item{slots}{a list of slots to be plugged into a \code{Chain} object}
}
\value{
a \code{Chain} object ready to be passed to \code{single_chain()}
}
\description{
Construct a \code{Chain} object, the
one argument to \code{single_chain(...)}.
}
\details{
If \code{list} is set, all other arguments will be dropped.
}

