% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/configs.R
\docType{class}
\name{Configs-class}
\alias{Configs-class}
\title{Class \code{Configs}}
\description{
A collection of MCMC control parameters.
}
\section{Slots}{

\describe{
\item{\code{burnin}}{MCMC burnin, the number of MCMC iterations to ignore at the beginning of each obj}

\item{\code{effects_update_beta}}{values of l for which to update the beta_{l, g} parameters. For debugging only.}

\item{\code{genes_return}}{Indices of genes whose parameter samples you want to return.
Applies to all gene-specific parameters except for the epsilons.}

\item{\code{genes_return_epsilon}}{Indices of genes g for which epsilon_{n, g} is updated/returned.}

\item{\code{iterations}}{Number of MCMC iterations after burnin for which selected parameter samples are kept.
Total MCMC iterations = burnin + thin * "iterations", and the whole "thin * iterations" portion
is used to calculate posterior means, mean squares, and probabilities.}

\item{\code{libraries_return}}{Indices of RNA-seq libraries whose parameter samples you want to return.
Currently moot because there are no library-specific parameters other than the epsilons, but that
could change in future versions of the package.}

\item{\code{libraries_return_epsilon}}{Indices of RNA-seq libraries n for which epsilon_{n, g} is updated/returned.
Applies to all library-specific parameters except for the epsilons.}

\item{\code{parameter_sets_return}}{Character vector naming the variables whose MCMC samples
you want to return}

\item{\code{parameter_sets_update}}{Character vector naming the variables to calculate/update
during the MCMC.}

\item{\code{priors}}{Names of the family of priors on the betas after integrating out the xi's.
Can be any value returned by \code{special_beta_priors()}. All other values will default to the normal prior.}

\item{\code{samplers}}{character string indicating the sampling algorithm}

\item{\code{thin}}{MCMC thinning interval. \code{thin = 1} means parameter samples will be saved for every iterations
after burnin. \code{thin = 10} means parameter samples will be saved every 10th iteration after burnin.
Total MCMC iterations = burnin + thin * "iterations", and the whole "thin * iterations" portion
is used to calculate posterior means, mean squares, and probabilities.}

\item{\code{verbose}}{Number of times to print out progress during burnin and the actual MCMC.
If \code{verbose} > 0, then progress messages will also print during setup and cleanup.}
}}

