% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{effect_sizes}
\alias{effect_sizes}
\title{Function \code{effect_sizes}}
\usage{
effect_sizes(obj)
}
\arguments{
\item{obj}{a \code{Chain} object or a list of \code{Chain} objects returned by \code{fbseq()}.}
}
\value{
a data frame of estimated effect sizes, with rows corresponding to genes and columns 
corresponding to the propositions in \code{Scenario(chain)@propositions}.
}
\description{
Computes the "effect sizes" of propositions involving contrasts.
}

