% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates}
\alias{estimates}
\title{Function \code{estimates}}
\usage{
estimates(chain, level = 0.95)
}
\arguments{
\item{chain}{a \code{Chain} object}

\item{level}{level of the credible intervals from 0 to 1}
}
\value{
A list of (1) "probs" a data frame of heterosis probabilities (with rows matching those of the
original count data), (2) "params" a data frame of estimates: posterior means and approximate credible intervals.
Parameters not updeted in the MCMC are excluded, and (3) "psrf" Gelman-Rubin potential scale reduction factors
(empty if not using Gelman diagnostics)
}
\description{
Extracts heterosis probabilities, posterior means, and estimated credible intervals
from a \code{Chain} object. The order of the rows respects the order of the rows/features/genes
in the original RNA-seq count data.
}

