% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chain_constructor.R
\name{fill_easy_gaps}
\alias{fill_easy_gaps}
\title{Function \code{fill_easy_gaps}}
\usage{
fill_easy_gaps(chain, counts, design)
}
\arguments{
\item{chain}{a \code{Chain} object}

\item{counts}{Matrix of RNA-seq read counts.}

\item{design}{Gene-specific design matrix. Must contain only 0's, 1's, and -1's.
Must have rows corresponding to colums/libraries in RNA-seq data and colums corresponding to
gene-specific variables.}
}
\value{
a \code{Chain} object
}
\description{
Set required slots in a \code{Chain} object.
Used internally.
}

