% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chain_constructor.R
\name{fill_easy_gaps}
\alias{fill_easy_gaps}
\title{Function \code{fill_easy_gaps}}
\usage{
fill_easy_gaps(chain, scenario)
}
\arguments{
\item{chain}{a \code{Chain} object}

\item{scenario}{an \code{Scenario} object
Must have rows corresponding to colums/libraries in RNA-seq data and colums corresponding to
sets of gene-specific variables.}
}
\value{
a \code{Chain} object
}
\description{
Set required slots in a \code{Chain} object.
Used internally.
}

