% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Function \code{generate_data}}
\usage{
generate_data(libraries = 12, genes = 35000, design = cbind(rep(1,
  libraries), rep(c(1, -1, 1), each = floor(libraries/3)), rep(c(-1, 1, 1), each
  = floor(libraries/3))), starts = Starts(nu = 10, omegaSquared = 0.01, pi =
  c(0.5, 0.5, 0.5), sigmaSquared = c(2.25, 0.25, 0.25), tau = 0.1, theta = c(5,
  0, 0)))
}
\arguments{
\item{libraries}{number of libraries/libraries in the data}

\item{genes}{number of genes/genes in the data}

\item{design}{Gene-specific design matrix.
Must have rows corresponding to colums/libraries in RNA-seq data and colums corresponding to
gene-specific variables.
Can be among "Laplace", "t", or "horseshoe". All other values will default to the normal prior.}

\item{starts}{A \code{Starts} object containing the necessary initialization constants
and hyperparameters.}
}
\value{
a list
\describe{
  \item{\code{counts}}{The generated RNA-seq counts matrix.}
  \item{\code{design}}{The design matrix for gene-specific effects.}
  \item{\code{truth}}{A \code{Starts} object containing the values used to simulate the data}
}
}
\description{
Simulates heterosis data from prespecified hyperparameters.
}

