% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chain_constructor.R
\name{plug_in_chain}
\alias{plug_in_chain}
\title{Function \code{plug_in_chain}}
\usage{
plug_in_chain(chain, configs, starts)
}
\arguments{
\item{chain}{a \code{Chain} object}

\item{configs}{A \code{Configs} object of MCMC control parameters.}

\item{starts}{A \code{Starts} object of model parameter starting values.}
}
\value{
a \code{Chain} object
}
\description{
Plug the slots of \code{Configs} and \code{Starts}
objects into the analogous slots of a \code{Chain} object. Used
internally
}

