% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{probs}
\alias{probs}
\title{Function \code{probs}}
\usage{
probs(obj)
}
\arguments{
\item{obj}{a \code{Chain} object or a list of \code{Chain} objects returned by \code{fbseq()}.}
}
\value{
a data frame of estimated posterior probabilities, with rows corresponding to genes and columns 
corresponding to the propositions in the relevant \code{Scenario} object.
}
\description{
Extracts estimated posterior probs from a \code{Chain} object.
or a list of \code{Chain} objects returned by \code{fbseq()}.
}

