% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrf.R
\name{psrf}
\alias{psrf}
\title{Function \code{psrf}}
\usage{
psrf(chain_list)
}
\arguments{
\item{chain_list}{list of \code{Chain} objects}
}
\value{
Gelman potential scale reduction factors on a list of parallel independent chains.
}
\description{
calculate all Gelman potential scale reduction factors on a list of parallel independent chains.
}

