% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samplers.R
\name{Samplers}
\alias{Samplers}
\title{Constructor for class \code{Samplers}}
\usage{
Samplers(obj = NULL, ...)
}
\arguments{
\item{obj}{a \code{Chain} or \code{list} object to get slots from.}

\item{...}{additional slots.}
}
\description{
Constructor for class \code{Samplers}
}
\details{
Precedence will be given to the \code{Chain} or \code{list} object over \code{...}.
}

