% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scenario.R
\docType{class}
\name{Scenario-class}
\alias{Scenario-class}
\title{Class \code{Scenario}}
\description{
A collection of MCMC control parameters.
}
\section{Slots}{

\describe{
\item{\code{conjunctions}}{List of numeric vectors of length J. Each numeric vector is a collection of
indices c_1, ..., c_J such that the posterior probability,
P(contrasts[[c_1]] * beta > value[c_1], ..., contrasts[[c_J]] * beta > value[c_J] | data),
is estimated by averging over Monte Carlo samples. This allows the user to "test"
multiple contrasts together. See the vignette for a more thorough explanation.}

\item{\code{contrasts}}{list of numeric vectors, each length \code{L}. Each vector
is a contrast of the \code{beta_{l, .}}'s of interest.}

\item{\code{counts}}{RNA-seq count data matrix with G rows (for genes) and N columns (for libraries/samples).}

\item{\code{design}}{Design matrix, with \code{N} rows and \code{L} columns.
Same format as in \code{edgeR::glmFit}, with RNA-seq libraries/samples as rows
and each \code{beta_{l, .}} as a column.}

\item{\code{values}}{Numeric vector of length \code{length(contrasts)}. Used int tests.}
}}

