% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{Scenario}
\alias{Scenario}
\title{Constructor for class \code{Scenario}}
\usage{
Scenario(obj = NULL, ...)
}
\arguments{
\item{obj}{a \code{Chain} or \code{list} object to get slots from.}

\item{...}{optional slot bounds.}
}
\value{
a \code{Scenario} object
}
\description{
Precedence will be given to the \code{Chain} or \code{list} object over \code{...}.
Elements passed with \code{...} must be named. For example, \code{Scenario(design = my_matrix)}.
}

