% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scenario_heterosis_model.R
\name{scenario_heterosis_model}
\alias{scenario_heterosis_model}
\title{Function \code{scenario_heterosis_model}}
\usage{
scenario_heterosis_model(genes = 35000, libraries = 16)
}
\arguments{
\item{genes}{number of genes/genes in the data}

\item{libraries}{number of libraries/libraries in the data}
}
\value{
a \code{Scenario} object for the heterosis problem
}
\description{
produces a generic heterosis scenario
}

