% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_heterosis_model.R
\name{scenario_heterosis_model}
\alias{scenario_heterosis_model}
\title{Function \code{scenario_heterosis_model}}
\usage{
scenario_heterosis_model(genes = 30000, libraries = 16, truth = Starts(nu
  = 3, tau = 0.01, sigmaSquared = c(1, 0.04, 0.03, 5e-04, 0.1), theta = c(3, 0,
  -0.007, -0.005, 0.008)))
}
\arguments{
\item{genes}{number of genes/genes in the data}

\item{libraries}{number of libraries/libraries in the data}

\item{truth}{\code{Starts} object of hyperparameters underlying simulation
all hyperparameters must be specified.}
}
\value{
a \code{Scenario} object for the heterosis problem
}
\description{
produces a generic heterosis scenario
}

