% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scenario_heterosis_model.R
\name{scenario_heterosis_model}
\alias{scenario_heterosis_model}
\title{Function \code{scenario_heterosis_model}}
\usage{
scenario_heterosis_model(genes = 30000, libraries = 16, truth = Starts(nu
  = 2.58, tau = 0.00677, sigmaSquared = c(1, 0.0447, 0.034, 0.000473, 0.0763),
  theta = c(3, 0.00224, -0.0073, -0.00451, 0.00794)))
}
\arguments{
\item{genes}{number of genes/genes in the data}

\item{libraries}{number of libraries/libraries in the data}

\item{truth}{\code{Starts} object of hyperparameters underlying simulation
all hyperparameters must be specified.}
}
\value{
a \code{Scenario} object for the heterosis problem
}
\description{
produces a generic heterosis scenario
}

