% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scenario_heterosis_model.R
\name{scenario_heterosis_model}
\alias{scenario_heterosis_model}
\title{Function \code{scenario_heterosis_model}}
\usage{
scenario_heterosis_model(genes = 35000, libraries = 16, truth = Starts(nu
  = 2.812224, omegaSquared = 0.006808205, tau = 0.006780517, sigmaSquared = c(1,
  0.03724313, 0.0324207, 0.0006229287, 0.06410533), theta = c(3, -0.005734982,
  -0.02541216, -0.004763663, -0.06341044)), priors = "Laplace")
}
\arguments{
\item{genes}{number of genes/genes in the data}

\item{libraries}{number of libraries/libraries in the data}

\item{truth}{\code{Starts} object of hyperparameters underlying simulation
all hyperparameters must be specified.}

\item{priors}{character string, either a member of \code{special_beta_priors()} or
another string for normal priors on the betas.}
}
\value{
a \code{Scenario} object for the heterosis problem
}
\description{
produces a generic heterosis scenario
}

