% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starts.R
\docType{class}
\name{Starts-class}
\alias{Starts-class}
\title{Class \code{Starts}}
\description{
Set MCMC starting values and model constants listed
by \code{help("Starts-class")}.
}
\section{Slots}{

\describe{
\item{\code{a}}{initialization constant}

\item{\code{b}}{initialization constant}

\item{\code{c}}{initialization constants}

\item{\code{d}}{initialization constant}

\item{\code{h}}{initialization constants}

\item{\code{k}}{initialization constants}

\item{\code{q}}{initialization constants}

\item{\code{r}}{initialization constants}

\item{\code{s}}{initialization constants}

\item{\code{beta}}{MCMC starting values}

\item{\code{epsilon}}{MCMC starting values}

\item{\code{gamma}}{MCMC starting values}

\item{\code{nu}}{MCMC starting values}

\item{\code{sigmaSquared}}{MCMC starting values}

\item{\code{tau}}{MCMC starting values}

\item{\code{theta}}{MCMC starting values}

\item{\code{xi}}{MCMC starting values}
}}

