% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/starts.R
\docType{class}
\name{Starts-class}
\alias{Starts-class}
\title{Class \code{Starts}}
\description{
A way for users to specify model parameter
starting values. Use only when creating new \code{obj}
objects from scratch.
}
\section{Slots}{

\describe{
\item{\code{aRho}}{initialization constant}

\item{\code{aGamma}}{initialization constant}

\item{\code{bRho}}{initialization constant}

\item{\code{bGamma}}{initialization constant}

\item{\code{c}}{initialization constants}

\item{\code{dRho}}{initialization constant}

\item{\code{dGamma}}{initialization constant}

\item{\code{h}}{log-scale normalization factors}

\item{\code{k}}{initialization constants}

\item{\code{r}}{initialization constants}

\item{\code{s}}{initialization constants}

\item{\code{beta}}{MCMC starting values}

\item{\code{epsilon}}{MCMC starting values}

\item{\code{gamma}}{MCMC starting values}

\item{\code{nuGamma}}{MCMC starting values}

\item{\code{nuRho}}{MCMC starting values}

\item{\code{rho}}{MCMC starting values}

\item{\code{sigmaSquared}}{MCMC starting values}

\item{\code{tauGamma}}{MCMC starting values}

\item{\code{tauRho}}{MCMC starting values}

\item{\code{theta}}{MCMC starting values}

\item{\code{xi}}{MCMC starting values}
}}

