% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starts.R
\name{Starts}
\alias{Starts}
\title{Constructor for class \code{Starts}}
\usage{
Starts(obj = NULL, ...)
}
\arguments{
\item{obj}{a \code{Chain} or \code{list} object to get slots from.}

\item{...}{additional slots.}
}
\description{
Precedence will be given to the \code{Chain} or \code{list} object over \code{...}.
}

