% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano}
\alias{volcano}
\title{Function \code{volcano}}
\usage{
volcano(obj)
}
\arguments{
\item{obj}{a \code{Chain} object or a list of \code{Chain} objects returned by \code{fbseq()}.}
}
\value{
a volcano plot of probabilities of propositions vs the effect sizes
}
\description{
Computes volcano plots by plotting probabilities of propositions vs the effect sizes
}

